de int a double;

+++ interficies/Cercle.java
@@ -2,17 +2,17 @@ package interficies;
 
 class Cercle extends interficies.Figura2D {
 
-    public Cercle(int base, int altura) {
+    public Cercle(double base, double altura) {
         super(base, altura);
     }
 
     @Override
     public void calcularArea() {
-    ;
+        ;
     }
 
     @Override
     public void calcularPerimetre() {
-    ;
+        ;
     }
 }
